Minor refactor: split huge runtimeoptimize.cpp -> constfold.cpp
